Report capped Warp themes across directories
